Move `InMemoryFeedStore` to its own file
