+++ HackrNews/Story Details Feature/StoryDetail.swift
@@ -33,8 +33,4 @@ public struct StoryDetail {
         self.comments = comments
         self.url = url
     }
-
-    public var hasBody: Bool {
-        text != nil
-    }
 }
